concat server-new-project-type & client-new-project-wizard into one unique project new-project-type-wizard
correcting GinModule annotation

updating README

cleaning client-new-project-wizard

updating examples parent pom + name
