--- examples/client-new-project-wizard/README.md
@@ -1,3 +1,0 @@
-# Project Wizard tutorial
-In this module you will find how to:
-- Extend the project wizard
