+++ examples/new-project-type-wizard/.codenvy/project.json
