+++ examples/new-project-type-wizard/README.md
@@ -1,0 +1,4 @@
+# Project Wizard tutorial
+In this module you will find how to:
+- Add a new project type
+- Extend the project wizard
