+++ examples/new-project-type-wizard/src/main/assembly/tutorial-archive.xml
