+++ examples/new-project-type-wizard/src/main/java/org/eclipse/che/examples/projectwizard/client/SampleProjectWizardRegistrar.java
