+++ examples/new-project-type-wizard/src/main/java/org/eclipse/che/examples/projectwizard/shared/ProjectAttributes.java
