+++ examples/new-project-type-wizard/src/main/resources/org/eclipse/che/examples/projectwizard/SampleProjectWizardExtension.gwt.xml
