--- examples/server-new-project-type/.codenvy/project.json
@@ -1,12 +1,0 @@
-{"type": "codenvy_extension",
-    "builders": {
-        "default": "maven"
-    },
-    "runners": {
-        "configs": {},
-        "default": "system:/sdk/tomcat7"
-    },
-    "attributes": {
-        "language": ["java"]
-    }
-}
