--- examples/server-new-project-type/README.md
@@ -1,3 +1,0 @@
-# Project Type tutorial
-In this module you will find how to:
-- Add a new project type
