--- examples/server-new-project-type/src/main/java/org/eclipse/che/examples/projecttype/shared/ProjectAttributes.java
@@ -1,18 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2015 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.examples.projecttype.shared;
-
-public interface ProjectAttributes {
-    String SAMPLE_PROJECT_TYPE_ID       = "sample";
-    String SAMPLE_PROJECT_TYPE_NAME     = "Sample Project Type Project";
-    String SAMPLE_PROJECT_TYPE_CATEGORY = "Sample Project Type";
-    String PROGRAMMING_LANGUAGE         = "java";
-}
