--- examples/server-new-project-type/src/main/resources/org/eclipse/che/examples/projecttype/SampleProjectTypeExtension.gwt.xml
@@ -1,18 +1,0 @@
-<!--
-
-    Copyright (c) 2012-2015 Codenvy, S.A.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Codenvy, S.A. - initial API and implementation
-
--->
-<module>
-    <inherits name="com.google.gwt.user.User"/>
-    <inherits name="org.eclipse.che.ide.Api"/>
-    <inherits name="com.google.gwt.inject.Inject"/>
-    <source path=""/>
-</module>
